content/base/test/test_xhr_forbidden_headers.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=308484
-->
<head>
  <title>Test for Bug 308484</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>        
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=308484">Mozilla Bug 308484</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Bug 308484 **/

var headers = [
  "aCCept-chaRset",
  "acCePt-eNcoDing",
  "aCcEsS-cOnTrOl-ReQuEsT-mEtHoD",
  "aCcEsS-cOnTrOl-ReQuEsT-hEaDeRs",
  "coNnEctIon",
  "coNtEnt-LEngth",
  "CoOKIe",
  "cOOkiE2",
  "cOntEnt-tRAnsFer-enCoDiNg",
  "DATE",
  "exPeCt",
  "hOSt",
  "keep-alive",
  "oRiGiN",
  "reFERer",
  "te",
  "trAiLer",
  "trANsfEr-eNcoDiNg",
  "uPGraDe",
  "user-AGENT",
  "viA",
  "pRoxy-",
  "sEc-",
  "proxy-fOobar",
  "sec-bAZbOx"
];
var i, request;

// Try setting headers in unprivileged context
request = new XMLHttpRequest();
request.open("GET", window.location.href);
for (i = 0; i < headers.length; i++)
  request.setRequestHeader(headers[i], "test" + i);

// Read out headers
netscape.security.PrivilegeManager.enablePrivilege("UniversalXPConnect");
var channel = request.channel.QueryInterface(Components.interfaces.nsIHttpChannel);
for (i = 0; i < headers.length; i++) {
  // Retrieving Content-Length will throw an exception
  var value = null;
  try {
    value = channel.getRequestHeader(headers[i]);
  }
  catch(e) {}

  isnot(value, "test" + i, "Setting " + headers[i] + " header in unprivileged context");
}

// Try setting headers in privileged context
request = new XMLHttpRequest();
request.open("GET", window.location.href);
for (i = 0; i < headers.length; i++)
  request.setRequestHeader(headers[i], "test" + i);

// Read out headers
var channel = request.channel.QueryInterface(Components.interfaces.nsIHttpChannel);
for (i = 0; i < headers.length; i++) {
  var value = channel.getRequestHeader(headers[i]);
  is(value, "test" + i, "Setting " + headers[i] + " header in privileged context");
}
</script>
</pre>
</body>
</html>
